Refactor folder opening
The previous logic would fail if the folder contained any record
without an original file.  This now gracefully ignores flawed files,
which should result in a more robust overall user experience.

Files without an original file will, from the perspective of SFTP,
simply not exist.

Issue #57
