+++ src/utils/index.ts
@@ -1,4 +1,7 @@
 export * from './generateAttributesForFile';
 export * from './generateDefaultAttributes';
+export * from './generateFileEntriesForFolders';
+export * from './generateFileEntriesForRecords';
 export * from './generateFileEntry';
 export * from './getLongname';
+export * from './getOriginalFileForRecord';
